requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==4.2.5
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-tinymce==3.6.1
djangorestframework==3.15.1
Faker==21.0.0
idna==3.7
Pillow==9.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
transliterate==1.10.2
typing_extensions==4.11.0
urllib3==2.2.1
